packages/extractors/pyproject.toml
[project]
name = "extractors"
version = "0.1.0"
requires-python = ">=3.13"
dependencies = [
    "anthropic>=0.96.0",
    "cyvcf2>=0.32.1",
    "pillow>=12.2.0",
    "pydantic>=2.0.0",
    "pytesseract>=0.3.13",
]

[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
where = [".."]
include = ["extractors*"]

[dependency-groups]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.24.0",
]

[tool.ruff]
line-length = 100
target-version = "py313"

[tool.ruff.lint]
select = ["E", "F", "I", "UP"]
ignore = ["E501"]
